chore: move color utils to src/
