+++ packages/celeste-vue/src/components/color-picker/color-swatch.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
+import { DEFAULT_SWATCH } from '@/utils/color';
 import tinycolor from 'tinycolor2';
 import { computed } from 'vue';
 import { defineColorModel } from './composable/use-color-model';
-import { DEFAULT_SWATCH } from './utils';
 
 const props = withDefaults(defineProps<ColorSwatchProps>(), {
   label: 'Recommended Colors',
